ref: FileInterface.php
